requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3